+++ piyavkin.anton/M1/main.cpp
@@ -23,6 +23,7 @@ int main(int argc, char* argv[])
     std::cerr << "The number of command line arguments does not meet the requirements\n";
     return 1;
   }
+
   size_t tries = 0;
   size_t seed = 0;
   try
@@ -42,8 +43,9 @@ int main(int argc, char* argv[])
     std::cerr << e.what() << '\n';
     return 2;
   }
-  long long r = 0;
-  int countThreads = 0;
+
+  double r = 0;
+  long countThreads = 0;
   std::minstd_rand gen(seed);
   while (std::cin >> r >> countThreads)
   {
@@ -58,10 +60,14 @@ int main(int argc, char* argv[])
       std::vector< std::pair< double, double > > points;
       points.reserve(tries);
       createPoints(gen, points, r, tries);
+      
       auto start = std::chrono::high_resolution_clock::now();
+      
       double square = getSquare(points, r, countThreads);
+      
       auto end = std::chrono::high_resolution_clock::now();
       auto mcs = std::chrono::duration_cast< std::chrono::microseconds >(end - start).count();
+      
       StreamGuard guard(std::cout);
       std::cout << std::fixed << std::setprecision(3);
       std::cout << static_cast< double >(mcs) / 1000 << ' ' << square << '\n';
